Petites corrections dans les repository (Id:Integer => Long) / Ajout de la gestion des exceptions pour les controllers / Ajout de la gestion d'une réservation d'une annonce

+++ src/main/java/dev/repository/AnnonceRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import dev.domain.Annonce;
 
-public interface AnnonceRepo extends JpaRepository<Annonce, Integer> {
+public interface AnnonceRepo extends JpaRepository<Annonce, Long> {
 
 }
