+++ src/main/java/dev/repository/VehiculeRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import dev.domain.Vehicule;
 
-public interface VehiculeRepo extends JpaRepository<Vehicule, Integer> {
+public interface VehiculeRepo extends JpaRepository<Vehicule, Long> {
 
 }
